requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pyserial==2.7
SQLAlchemy==0.9.8
Werkzeug==0.10.4
